added insert_class handler and incomplete preferences
